# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Flask API for Time Series Data",
    "description": "This API allows users to upload CSV files and retrieve time series data.",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "schemes": ["http"],
  "paths": {
    "/upload": {
      "post": {
        "summary": "Upload a CSV file",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "required": true,
            "type": "file",
            "description": "CSV file to upload"
          }
        ],
        "responses": {
          "200": {
            "description": "File uploaded successfully"
          },
          "400": {
            "description": "Invalid file format or missing file"
          }
        }
      }
    },
    "/get_timeseries": {
      "get": {
        "summary": "Retrieve 10 random time series data points",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Name of the uploaded CSV file"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns 10 time series data points",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "400": {
            "description": "Filename missing or not enough data points"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/get_forecast": {
      "get": {
        "summary": "Create 3 forecast time series data points based on previous 10 data points.",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Name of the uploaded CSV file."
          },
          {
            "name": "start_idx",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Last index of the time series response."
          }
        ],
        "responses": {
          "200": {
            "description": "Returns 3 forcasted time series data points",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "400": {
            "description": "Filename missing or no starting index."
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/all_uploads": {
      "get": {
        "summary": "Queries the Database and retrieves all files uploaded.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns database files.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "400": {
            "description": "No files uploaded in the DB."
          }
        }
      }
    }

  }
}